+++ integrations/cli/pixeleye/src/booth.ts
@@ -25,7 +25,6 @@ program
       endpoint: options.endpoint,
       token: token.substring(1, token.length - 1),
       buildID: buildID.substring(1, buildID.length - 1),
-      domEnvironment: options.domEnvironment,
     }).catch((err) => {
       program.error(err);
     });