+++ integrations/cli/pixeleye/src/commands.ts
@@ -1,7 +1,13 @@
 import chalk from "chalk";
 import { Command } from "commander";
 import { loadAndMergeConfig } from "./config-loader";
-import { execHandler, ping, storybook, uploadHandler } from "./handlers";
+import {
+  execHandler,
+  ping,
+  snapFileHandler,
+  storybook,
+  uploadHandler,
+} from "./handlers";
 import { setLogLevel } from "@pixeleye/cli-logger";
 
 export const program = new Command();
@@ -18,6 +24,8 @@ export const optionMap = {
   w: "wait",
   v: "verbose",
   c: "config",
+  sm: "sitemaps",
+  u: "urls",
 } as const;
 
 const configOption = (name: string) =>
@@ -90,5 +98,29 @@ apiOptions("upload")
   .description("Upload screenshots to pixeleye")
   .hook("preAction", loadAndMergeConfig)
   .action(uploadHandler);
+
+apiOptions("snapshot")
+  .argument(
+    "[files...]",
+    "File globs for pixeleye snapshot definitions. You can also declare these in your config file"
+  )
+  .option(
+    "-sm, --sitemaps <sitemaps...>",
+    "Sitemap URLs to snapshot, e.g. https://example.com/sitemap.xml. If a url has already been listed in a file], the file definition will take precedence."
+  )
+  .option(
+    "-u, --urls <urls...>",
+    "URLs to snapshot, e.g. https://example.com. If a url has already been listed in a file, the file definition will take precedence."
+  )
+  .option(
+    "-w, --wait [wait]",
+    "Wait for build results, outputting them once finished processing",
+    "false"
+  )
+  .description(
+    "Upload screenshots to pixeleye from a list of URL files passed in or defined in your config file"
+  )
+  .hook("preAction", loadAndMergeConfig)
+  .action(snapFileHandler);
 
 export default program.parse(process.argv);