+++ integrations/cli/pixeleye/src/handlers/index.ts
@@ -2,3 +2,4 @@ export * from "./ping";
 export * from "./storybook";
 export * from "./exec";
 export * from "./upload";
+export * from "./snapFile";