+++ integrations/cli/cli-booth/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @pixeleye/cli-booth
 
+## 0.4.13
+
+### Patch Changes
+
+- Updated dependencies [[`57d5bf2`](https://github.com/pixeleye-io/pixeleye/commit/57d5bf28839bb7ebc450155312c0c4ae9a31627b)]:
+  - @pixeleye/cli-config@0.5.0
+  - @pixeleye/cli-capture@0.4.7
+
+## 0.4.12
+
+### Patch Changes
+
+- [#248](https://github.com/pixeleye-io/pixeleye/pull/248) [`c9cef90`](https://github.com/pixeleye-io/pixeleye/commit/c9cef900d102ba61a673cb7c9e06fe57a3f62999) Thanks [@AlfieJones](https://github.com/AlfieJones)! - Added wait for selector option
+
+- Updated dependencies [[`4730ec0`](https://github.com/pixeleye-io/pixeleye/commit/4730ec067e1e9ccb1ffd1232c0f14d567b784c23), [`c9cef90`](https://github.com/pixeleye-io/pixeleye/commit/c9cef900d102ba61a673cb7c9e06fe57a3f62999), [`c9cef90`](https://github.com/pixeleye-io/pixeleye/commit/c9cef900d102ba61a673cb7c9e06fe57a3f62999)]:
+  - @pixeleye/cli-capture@0.4.7
+
 ## 0.4.11
 
 ### Patch Changes