+++ integrations/cli/cli-capture/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pixeleye/cli-capture
 
+## 0.4.7
+
+### Patch Changes
+
+- [#248](https://github.com/pixeleye-io/pixeleye/pull/248) [`4730ec0`](https://github.com/pixeleye-io/pixeleye/commit/4730ec067e1e9ccb1ffd1232c0f14d567b784c23) Thanks [@AlfieJones](https://github.com/AlfieJones)! - Added retry ability
+
+- [#248](https://github.com/pixeleye-io/pixeleye/pull/248) [`c9cef90`](https://github.com/pixeleye-io/pixeleye/commit/c9cef900d102ba61a673cb7c9e06fe57a3f62999) Thanks [@AlfieJones](https://github.com/AlfieJones)! - Added wait for selector option
+
+- [#248](https://github.com/pixeleye-io/pixeleye/pull/248) [`c9cef90`](https://github.com/pixeleye-io/pixeleye/commit/c9cef900d102ba61a673cb7c9e06fe57a3f62999) Thanks [@AlfieJones](https://github.com/AlfieJones)! - Added wait for font to load
+
 ## 0.4.6
 
 ### Patch Changes